+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Improvements
 
+- PR #2109 added name to Python column classes
 - PR #1947 Cleanup serialization code
 - PR #2125 More aggregate in java API
 - PR #2131 Chunk rows logic added to csv_writer
@@ -22,6 +23,7 @@
 - PR #2151 Fixes bug in cudf::copy_range where null_count was invalid
 - PR #2139 matching to pandas describe behavior & fixing nan values issue
 - PR #2154 CSV Reader: Fix bools misdetected as strings dtype
+- PR #2178 Fix bug in rolling bindings where a view of an ephemeral column was being taken
 
 
 # cuDF 0.8.0 (27 June 2019)